Added spin.js spinner as activity indicator.
Previously this widget called `showLoader` on the map which may or may not have worked and may or may not have been seen by the user. Moving the activity indicator to be within the widget will make it more visible.

Also, [spin.js](http://fgnass.github.io/spin.js/) is super cool.